+++ examples/clip-volt/src/background/store/background.state.ts
@@ -1,5 +1,5 @@
-import { State } from '@hexajs/core';
-import { InjectableContext } from '@hexajs/common';
+import { State } from '@hexajs-dev/core';
+import { InjectableContext } from '@hexajs-dev/common';
 import { BackgroundState, ConfigReducer, ClipsReducer } from './background.reducer';
 
 @State<BackgroundState>({