+++ examples/hexa-grayscale/src/content/content.ts
@@ -1,5 +1,5 @@
-import { Content, ContentRunAt } from '@hexajs/core';
-import { OnInit, OnDestroy } from '@hexajs/common';
+import { Content, ContentRunAt } from '@hexajs-dev/core';
+import { OnInit, OnDestroy } from '@hexajs-dev/common';
 import { GrayscaleUiService } from './ui/grayscale-ui.service';
 
 @Content({ matches: ['<all_urls>'], runAt: ContentRunAt.DocumentIdle })