+++ examples/smart-clipper/ui/devtools/src/App.tsx
@@ -1,6 +1,6 @@
-﻿import { useEffect, useMemo, useState } from 'react';
-import { inject } from '@hexajs/common';
-import { RuntimePort } from '@hexajs/ports';
+import { useEffect, useMemo, useState } from 'react';
+import { inject } from '@hexajs-dev/common';
+import { RuntimePort } from '@hexajs-dev/ports';
 import { devtoolsHandlesApi } from '@contract/api';
 import { DevtoolsSyncThemeMessage } from '@contract/messages/messages';
 import smartClipperLogoUrl from '../../../src/assets/smart-clipper.logo.svg';